decouple WeightInfo trait from the weights module
